requirements.txt
asgiref==3.4.1
black==21.12b0
click==8.0.3
colorama==0.4.4
Django==4.0
djangorestframework==3.13.1
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pytz==2021.3
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
tzdata==2021.5
whitenoise==5.3.0
